+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet
+### Added
+
+- Basic VVC support with vvcC box and VvcDecoderConfigurationRecord
 
 ## [0.49.0] - 2025-06-26